[ELYEE-101] Adjust so we also have a global Policy instance to manage / fall back on.
