+++ authorization/src/main/java/org/wildfly/security/authz/jacc/ElytronPolicyFactory.java
@@ -15,25 +15,37 @@ import jakarta.security.jacc.PolicyFactory;
  * The Elytron {@code PolicyFactory} implementation.
  *
  * This is a very simple implementation with no Policy instantiation
- * as we will be expecting that to be handled by {@code PolicyRegistration}.
+ * as we will be expecting that to be handled by {@code AuthorizationRegistration}.
  *
  * @author <a href="mailto:darran.lofthouse@jboss.com">Darran Lofthouse</a>
  */
 public class ElytronPolicyFactory extends PolicyFactory {
 
-    private Map<String, Policy> contextPolicyMap = new ConcurrentHashMap<>();
+    private volatile Policy defaultPolicy;
+    private final Map<String, Policy> contextPolicyMap = new ConcurrentHashMap<>();
+
+    public ElytronPolicyFactory(final Policy defaultPolicy) {
+        this.defaultPolicy = defaultPolicy;
+    }
 
     @Override
     public Policy getPolicy(String contextId) {
-        return contextPolicyMap.get(contextId);
+        Policy policy = contextPolicyMap.get(contextId);
+
+        return policy != null ? policy : defaultPolicy;
     }
 
     @Override
     public void setPolicy(String contextId, Policy policy) {
-        if (policy == null) {
-            contextPolicyMap.remove(contextId);
+        if (contextId != null) {
+            if (policy == null) {
+                contextPolicyMap.remove(contextId);
+            } else {
+                contextPolicyMap.put(contextId, policy);
+            }
         } else {
-            contextPolicyMap.put(contextId, policy);
+            // If the contextId is null we are setting the default.
+            defaultPolicy = policy;
         }
     }
